% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addRandomizedPopulationData}
\alias{addRandomizedPopulationData}
\title{Add data from population studies to dataset.
Generate randomized data points within the given measurement
interval for the data, i.e. use mean and sd/range for x and y
to create n data points.
The data is weighted only a fraction of the individual data in
the regression but the information is provided for the fit curves.}
\usage{
addRandomizedPopulationData(data, newdata)
}
\description{
Add data from population studies to dataset.
Generate randomized data points within the given measurement
interval for the data, i.e. use mean and sd/range for x and y
to create n data points.
The data is weighted only a fraction of the individual data in
the regression but the information is provided for the fit curves.
}

