% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_half_max_time}
\alias{calc_half_max_time}
\title{Calculate half maximal times from the processed data}
\usage{
calc_half_max_time(processed, t_peak, t_end)
}
\description{
Preprocessing of the half maximal time from the response curves, i.e.
how fast is the half maximum reached after galactose challenge.
}

