% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEC_functions.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{Calculate quantiles for the data.}
\usage{
calc_quantiles(data, q.values = c(0.025, 0.25, 0.5, 0.75, 0.975))
}
\description{
Calculate quantiles for the data.
}

