% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate_f_Rbase}
\alias{calculate_f_Rbase}
\title{If the galactose elimination was not calculated based
on periportal - perivenious concentration differences, it is necessary to
correct the value for basal galactose removal by extraheptic tissues (~3%).
This has large effects on the clearance calculation.
Based on the Keiding1988 data this basal removal can be estimated assuming
similar removal kinetics for the extrahepatic tissues than for the liver
(also cleared by galactokinase)}
\usage{
calculate_f_Rbase()
}
\description{
If the galactose elimination was not calculated based
on periportal - perivenious concentration differences, it is necessary to
correct the value for basal galactose removal by extraheptic tissues (~3%).
This has large effects on the clearance calculation.
Based on the Keiding1988 data this basal removal can be estimated assuming
similar removal kinetics for the extrahepatic tissues than for the liver
(also cleared by galactokinase)
}

