% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate_GEC}
\alias{calculate_GEC}
\title{Calculates GEC for given vectors of liver volume and blood flow.}
\usage{
calculate_GEC(GEC_f, volLiver, flowLiver, f_tissue = 0.8)
}
\description{
Calculates GEC for given vectors of liver volume and blood flow.
}

