% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate_Rbase}
\alias{calculate_Rbase}
\title{Calculate the actual basal rate for given galactos concentration
Returns galactose remvoal in [mmol/min].}
\usage{
calculate_Rbase(gal, f = calculate_f_Rbase())
}
\description{
Calculate the actual basal rate for given galactos concentration
Returns galactose remvoal in [mmol/min].
}

