% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PEOPLE_predict_functions.R
\name{calculateBSA}
\alias{calculateBSA}
\title{Calculate body surface area (BSA) [m^2] from bodyweight [kg] and height [cm].}
\usage{
calculateBSA(bodyweight_kg, height_cm)
}
\description{
Using DuBois formula
}

