% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{calculateProbabilitiesForVariables}
\alias{calculateProbabilitiesForVariables}
\title{Calculate the single variable probabilies based on ECDF}
\usage{
calculateProbabilitiesForVariables(pars, ecdf.list)
}
\description{
Calculate the single variable probabilies based on ECDF
}

