% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateSampleProbability}
\alias{calculateSampleProbability}
\title{Calculate the overall probability of the sample under the assumption
of statistical independence.}
\usage{
calculateSampleProbability(pars, ps.var)
}
\description{
TODO: make sure it is valid
}

