% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Figure of correlation data}
\usage{
correlation_plot(data, m1, main, xname, yname, xlab, ylab, xlim, ylim,
  create_plots = F)
}
\description{
Figure of correlation data
}

