% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.settings}
\alias{create.settings}
\title{Create the settings structure.}
\usage{
create.settings(dir = "/home/mkoenig", simulator = "ROADRUNNER")
}
\description{
Encodes the main dirs and settings necessary.
Simulator can be either ROADRUNNER or COPASI.
}

