% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PEOPLE_predict_functions.R
\name{create_all_people}
\alias{create_all_people}
\title{Create people from the raw data files like 'hei1999'.}
\usage{
create_all_people(names)
}
\description{
Create people from the raw data files like 'hei1999'.
}

