% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PEOPLE_predict_functions.R
\name{create_people_from_raw}
\alias{create_people_from_raw}
\title{Create the people data frame used for prediction}
\usage{
create_people_from_raw(name)
}
\description{
Create the people data frame used for prediction
}

