% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createClearanceDataFrame}
\alias{createClearanceDataFrame}
\title{Calculate the clearance parameters and data.frame.}
\usage{
createClearanceDataFrame(folder, t_peak = 2000, t_end = 10000)
}
\arguments{
\item{folder}{preprocessed folder for analysis}
}
\value{
data.frame with clearance parameters
}
\description{
Calculate the clearance parameters and data.frame.
}

