% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createDataFromHistogramm}
\alias{createDataFromHistogramm}
\title{Generates a possible data vector from given histogramm.
Necessary for fitting distributions to histogramm data.
Count data points are generated in the middle of every bin.}
\usage{
createDataFromHistogramm(dset)
}
\arguments{
\item{dset}{histogram dataset}
}
\value{
data frame with x as data
}
\description{
Generates a possible data vector from given histogramm.
Necessary for fitting distributions to histogramm data.
Count data points are generated in the middle of every bin.
}

