\name{createDataMatrices}
\alias{createDataMatrices}
\title{Convert list structure into data matrix}
\usage{
createDataMatrices(dir, datalist)
}
\arguments{
  \item{datalist}{list of data matrices}

  \item{compounds}{which compounds to take}
}
\value{
matrix of pppv data
}
\description{
Convert list structure into data matrix
}

