% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createGalactoseClearanceDataFrame}
\alias{createGalactoseClearanceDataFrame}
\title{Calculate the clearance parameters and data.frame.
Necessary to provide the peak and simulation end time for calculation.
Here additional values from the simulation curves are added.}
\usage{
createGalactoseClearanceDataFrame(t_peak = 2000, t_end = 10000)
}
\arguments{
\item{folder}{preprocessed folder for analysis}
}
\value{
data.frame with clearance parameters
}
\description{
Calculate the clearance parameters and data.frame.
Necessary to provide the peak and simulation end time for calculation.
Here additional values from the simulation curves are added.
}

