% Generated by roxygen2 (4.0.2): do not edit by hand
\name{disease_predictor}
\alias{disease_predictor}
\title{Predict liver disease based on multiscale-model predictions of GEC.}
\usage{
disease_predictor(GEC_exp, GEC, q = 0.05)
}
\description{
The GEC multiscale classifier works via comparing the predicted GEC range
with the experimental value and uses the location of the experimental
data relative to the theoretical prediction based on the anthropomorphic
features for evaluation.
}

