\name{extendParameterStructure}
\alias{extendParameterStructure}
\title{Create extended data frame with the calculated values}
\usage{
extendParameterStructure(pars, fixed_ps, model)
}
\arguments{
  \item{pars}{parameter structure}
}
\description{
Create extended data frame with the calculated values
}

