% Generated by roxygen2 (4.0.2): do not edit by hand
\name{f_bootstrap}
\alias{f_bootstrap}
\title{Bootstrap of the calculation function.
The number of samples in the bootstrap corresponds to the available samples.}
\usage{
f_bootstrap(dset, funct, B = 1000)
}
\description{
Bootstrap of the calculation function.
The number of samples in the bootstrap corresponds to the available samples.
}

