% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GAMLSS_functions.R
\name{f_d.flowLiver.c}
\alias{f_d.flowLiver.c}
\title{Calculate the combined liver bloodflow distribution for person.}
\usage{
f_d.flowLiver.c(x, pars)
}
\description{
Calculate the combined liver bloodflow distribution for person.
}

