% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GAMLSS_functions.R
\name{f_d.flowLiverkg.pars}
\alias{f_d.flowLiverkg.pars}
\title{Calculate the liver bloodflow per bodyweight distributions for person.}
\usage{
f_d.flowLiverkg.pars(person, sex_split)
}
\description{
Calculate the liver bloodflow per bodyweight distributions for person.
}

