% Generated by roxygen2 (4.0.2): do not edit by hand
\name{f_d.parameters}
\alias{f_d.parameters}
\title{Factory to create various probability densities from individual models.}
\usage{
f_d.parameters(models, xname, person, sex_split = TRUE)
}
\description{
The prediction is computational expensive. The parameters for the distribution
have to be predicted exactely once per given person information.
}

