% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GAMLSS_functions.R
\name{f_d.rejection_sample}
\alias{f_d.rejection_sample}
\title{Rejection sampling from given distribution.}
\usage{
f_d.rejection_sample(f_d, Nsim, interval)
}
\description{
Sampling from distributions via rejection sampling of good guess of
probability distribution.
}

