% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GAMLSS_functions.R
\name{f_d.volLiver.c}
\alias{f_d.volLiver.c}
\title{Calculate the combined liver volume distribution for person.}
\usage{
f_d.volLiver.c(x, pars)
}
\description{
Calculate the combined liver volume distribution for person.
}

