% Generated by roxygen2 (4.0.1): do not edit by hand
\name{f_d.volLiverkg.c}
\alias{f_d.volLiverkg.c}
\title{Calculate the combined liver volume per bodyweight distribution for person.}
\usage{
f_d.volLiverkg.c(x, pars)
}
\description{
Calculate the combined liver volume per bodyweight distribution for person.
}

