% Generated by roxygen2 (4.0.2): do not edit by hand
\name{f_d.volLiverkg.pars}
\alias{f_d.volLiverkg.pars}
\title{Calculate the liver volume per bodyweight distributions for person.}
\usage{
f_d.volLiverkg.pars(person)
}
\description{
Calculate the liver volume per bodyweight distributions for person.
}

