% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{gender.cols}
\alias{gender.cols}
\title{Returns the colors for the different gender settings in data.}
\usage{
gender.cols()
}
\value{
gender levels and colors
}
\description{
Returns the colors for the different gender settings in data.
}

