% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEC_settings.R
\name{get_age_GE_folders}
\alias{get_age_GE_folders}
\title{Definition of GE curves for age dependency.
Depends on model. Update this folder when model is updated.}
\usage{
get_age_GE_folders()
}
\description{
Definition of GE curves for age dependency.
Depends on model. Update this folder when model is updated.
}

