% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEC_settings.R
\name{get_vmax_GE_folders}
\alias{get_vmax_GE_folders}
\title{Definition of GE curves for Vmax dependency.}
\usage{
get_vmax_GE_folders()
}
\description{
Definition of GE curves for Vmax dependency.
}

