\name{getBreakPointsFromMidpoints}
\alias{getBreakPointsFromMidpoints}
\title{Calculatesbreakpoints for histogram data with equidistant midpoints.}
\usage{
getBreakPointsFromMidpoints(midpoints)
}
\arguments{
  \item{midpoints}{vector of midpoints of bins}
}
\value{
vector of breakpoints
}
\description{
Calculatesbreakpoints for histogram data with equidistant
midpoints.
}

