% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFixedParameters}
\alias{getFixedParameters}
\title{Get fixed parameters in the model.}
\usage{
getFixedParameters(pars, all_ps)
}
\arguments{
\item{pars}{parameter structure}

\item{all_ps}{parameters in model}
}
\description{
Fixed parameters are all parameters which are listed in
the parameter names from the parameter structure.
}

