\name{getMaximaFromDilutionData}
\alias{getMaximaFromDilutionData}
\title{Get the maximal values of the dilution data}
\usage{
getMaximaFromDilutionData(data, compounds = compounds)
}
\arguments{
  \item{data}{experimental dataset}

  \item{correctTime}{should the time be corrected}
}
\description{
Get the maximal values of the dilution data
}

