% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{getParameterNames}
\alias{getParameterNames}
\title{Get parameter names, i.e columns which are not keywords.}
\usage{
getParameterNames(pars)
}
\value{
vector of parameter names
}
\description{
Uses the defined pars.keywords to see which names are parameters and
which are keywords.
}

