% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{getParameterTypes}
\alias{getParameterTypes}
\title{Get the different parameter types (fixed and variable)}
\usage{
getParameterTypes(pars)
}
\description{
Get the different parameter types (fixed and variable)
}

