% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRangeType}
\alias{getRangeType}
\title{Check if Sd or Range for x and y based on which fields are available
in the dataset}
\usage{
getRangeType(dat, xname, yname)
}
\description{
Check if Sd or Range for x and y based on which fields are available
in the dataset
}

