\name{getSimulationFileFromSimulationId}
\alias{getSimulationFileFromSimulationId}
\title{Get integration timecourse file for simulation id.}
\usage{
getSimulationFileFromSimulationId(simId)
}
\arguments{
  \item{simId}{simulation identifier}
}
\value{
filename for the integration file
}
\description{
Get integration timecourse file for simulation id.
}

