\name{getTimeFromMIMAT}
\alias{getTimeFromMIMAT}
\title{Reads the time vector from the simulation matrix list.}
\usage{
getTimeFromMIMAT(MI.mat)
}
\arguments{
  \item{MI.mat}{List of simulation matrixes}
}
\value{
time vector
}
\description{
Reads the time vector from the simulation matrix list.
}

