\name{getTimeFromPPMAT}
\alias{getTimeFromPPMAT}
\title{Reads the time vector from the simulation matrix list.}
\usage{
getTimeFromPPMAT(preprocess.mat)
}
\arguments{
  \item{MI.mat}{List of simulation matrixes}
}
\value{
time vector
}
\description{
Reads the time vector from the simulation matrix list.
}

