\name{getTimeFromPreprocessMatrix}
\alias{getTimeFromPreprocessMatrix}
\title{Reads the time vector from the simulation matrix list.}
\usage{
getTimeFromPreprocessMatrix(preprocess.mat)
}
\arguments{
  \item{MI.mat}{List of simulation matrixes}
}
\value{
time vector
}
\description{
Reads the time vector from the simulation matrix list.
}

