% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getVariableParameters}
\alias{getVariableParameters}
\title{Get parameters which are varied.}
\usage{
getVariableParameters(pars, all_ps)
}
\arguments{
\item{pars}{parameter structure}

\item{all_ps}{parameters in model}
}
\description{
Get parameters which are varied.
}

