% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PEOPLE_predict_functions.R
\name{individual_GEC_box}
\alias{individual_GEC_box}
\title{Boxplot of the GEC distribution.}
\usage{
individual_GEC_box(person, data, gec_exp = NA)
}
\description{
Boxplot of the GEC distribution.
}

