% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POPULATION_predict_functions.R
\name{individual_plot}
\alias{individual_plot}
\title{Full combined plot of the information.}
\usage{
individual_plot(person, vol, flow, data)
}
\description{
Full combined plot of the information.
}

