% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POPULATION_predict_functions.R
\name{individual_vol_flow_plot}
\alias{individual_vol_flow_plot}
\title{Scatterplot of liver volumes & blood flows.}
\usage{
individual_vol_flow_plot(person, vol, flow, data)
}
\description{
Scatterplot of liver volumes & blood flows.
}

