% Generated by roxygen2 (4.0.2): do not edit by hand
\name{integrate_GE_folders}
\alias{integrate_GE_folders}
\title{Preprocess set of data.frame of GE folders.}
\usage{
integrate_GE_folders(df_folders, factors = c("f_flow", "gal_challenge"),
  force = FALSE, tend = 10000, f_tissue = 0.8, vol_liver = 1500)
}
\description{
Preprocess set of data.frame of GE folders.
}

