% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{lim}
\alias{lim}
\title{Axis limits for given name.}
\format{\preformatted{List of 12
 $ age        : num [1:2] 0 100
 $ bodyweight : num [1:2] 0 135
 $ height     : num [1:2] 0 220
 $ BSA        : num [1:2] 0 3.1
 $ GEC        : num [1:2] 0 5
 $ GECkg      : num [1:2] 0 0.1
 $ volLiver   : num [1:2] 0 3200
 $ volLiverkg : num [1:2] 0 90
 $ volLiverBSA: num [1:2] 0 3200
 $ flowLiver  : num [1:2] 0 2800
 $ flowLiverkg: num [1:2] 0 90
 $ perfusion  : num [1:2] 0 2
}}
\usage{
lim
}
\value{
list of axis limits
}
\description{
Axis limits for given name.
}
\keyword{datasets}

