% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_models_for_prediction}
\alias{load_models_for_prediction}
\title{Load GAMLSS single correlation models.}
\usage{
load_models_for_prediction(dir = file.path(ma.settings$dir.base, "results",
  "gamlss"))
}
\description{
Load GAMLSS single correlation models.
}

