% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_shiny_fit_models}
\alias{load_shiny_fit_models}
\title{Load shiny fit models.}
\usage{
load_shiny_fit_models(file = file.path(ma.settings$dir.shiny,
  "fit_models.Rdata"))
}
\description{
Load the subset of GAMLSS models necessary for individual predictions.
}

