\name{loadParameterFile}
\alias{loadParameterFile}
\title{Read theparameter file into a data.frame.}
\usage{
loadParameterFile(file)
}
\arguments{
  \item{file}{parameter csv file for loading}
}
\value{
data.frame of parameters
}
\description{
This function reads the parameter file in CSV format into a
data.frame.
}

