% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{loadRawData}
\alias{loadRawData}
\title{Load the processed raw data for the given name.}
\usage{
loadRawData(name, dir = getRawDir())
}
\description{
This loads the the experimental datasets.
example: loadRawData('duc1979')
}

