% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBML_tools.R
\name{loadSBMLModel}
\alias{loadSBMLModel}
\title{Load SBML model from file.}
\usage{
loadSBMLModel(filename)
}
\arguments{
\item{filename}{Filename of the SBML file}
}
\value{
SBML model structure
}
\description{
Load SBML model from file.
}

