% Generated by roxygen2 (4.0.2): do not edit by hand
\name{m_split_sample}
\alias{m_split_sample}
\title{Split sample calculation of model fitting based on data and formula}
\usage{
m_split_sample(df, formula, B = 100)
}
\description{
Split sample calculation of model fitting based on data and formula
}

