\name{MultiscaleAnalysis-package}
\alias{MultiscaleAnalysis-package}
\alias{MultiscaleAnalysis}
\docType{package}
\title{
Analysis of multiscale model of liver 
}
\description{
Fit of parameter distributions for input, analysis of multiple indicator dilution curves,
analyis of galactose elimination (GEC), graphs for results.

}
\details{
\tabular{ll}{
Package: \tab MultiscaleAnalysis\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-15\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Matthias Koenig

Maintainer: Matthias Koenig <konigmatt@googlemail.com>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
