% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{pars.keywords}
\alias{pars.keywords}
\title{Reserved keywords which are not parameters.}
\format{\preformatted{ chr [1:4] "status" "duration" "core" "sim"
}}
\usage{
pars.keywords
}
\description{
Reserved keywords which are not parameters.
}
\keyword{datasets}

