% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_compound_curves}
\alias{plot_compound_curves}
\title{Plots individual multiple indicator dilution curves.}
\usage{
plot_compound_curves(time, data, name, weights, col = rgb(0.5, 0.5, 0.5, alpha
  = 0.1))
}
\description{
Data is a data matrix with simulations in columns and timepoints in rows.
Column ids correspond to the simulation identifiers, row ids to the timepoints.
}

