% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_compound_mean}
\alias{plot_compound_mean}
\title{Plot the mean data.}
\usage{
plot_compound_mean(time, data, weights, col)
}
\description{
Plots the mean and standard deviation for time courses. Uses weighted calculation
if the weights are provided.
Other available functions are rowMins, rowMaxs & rowQuantiles.
}

