% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{plot_correlation_data}
\alias{plot_correlation_data}
\title{Figure of correlation data}
\usage{
plot_correlation_data(data, m1, xname, yname, create_plots = F)
}
\description{
Figure of correlation data
}

