% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_single_compound}
\alias{plot_single_compound}
\title{Plot a single component of the mulitple indicator dilution curves.}
\usage{
plot_single_compound(time, data, name, col = "black", ylim = c(0, 0.2),
  xlim = c(0, 30), weights = NULL, ccols = NULL, meanData = TRUE)
}
\description{
Plot a single component of the mulitple indicator dilution curves.
}

