% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCompoundMean}
\alias{plotCompoundMean}
\title{Plot the mean data.}
\usage{
plotCompoundMean(time, data, weights, col)
}
\description{
Plots the mean and variance for time courses. Uses weighted calculation
if the weights are provided.
}

