\name{plotDistribution}
\alias{plotDistribution}
\title{Plot the density distribution.}
\usage{
plotDistribution(p.gen, name, maxvalue, N = 1000)
}
\arguments{
  \item{p.gen}{information about the parameters}

  \item{name}{of the parameter}

  \item{maxvalue}{for calculation of distribution}
}
\description{
Plot the density distribution.
}

