% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotFitHistogram}
\alias{plotFitHistogram}
\title{Plot the histogramm of the data used for fitting the distributions.}
\usage{
plotFitHistogram(p.gen, name, data, midpoints, col = rgb(1, 0, 0, 0.25))
}
\description{
Plot the histogramm of the data used for fitting the distributions.
}

