\name{plotHistogramForAllParameters}
\alias{plotHistogramForAllParameters}
\title{Creates histogram for all parameters.}
\usage{
plotHistogramForAllParameters(pars, all_pars = FALSE)
}
\arguments{
  \item{pars}{parameter data structure.}
}
\description{
This function generates all the histograms for the
parameters in the given parameter structure.
}

