% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotMultipleIndicatorMean}
\alias{plotMultipleIndicatorMean}
\title{All MultipleIndicator curves in one plot}
\usage{
plotMultipleIndicatorMean(time, preprocess.mat, weights,
  create_plot_files = F, xlim = c(0, 20), ylim = c(0, 1.5))
}
\description{
All MultipleIndicator curves in one plot
}

