\name{plotParameterHistogramFull}
\alias{plotParameterHistogramFull}
\title{Creates histogram for all parameters.}
\usage{
plotParameterHistogramFull(pars, all_pars = FALSE)
}
\arguments{
  \item{pars}{parameter data structure.}
}
\description{
This function generates all the histograms for the
parameters in the given parameter structure.
}

