\name{plotWeighted}
\alias{plotWeighted}
\title{Plot the weighted parameter values with underlying distribution.}
\usage{
plotWeighted(pars, p.gen, name)
}
\arguments{
  \item{pars}{parameter structure}

  \item{p.gen}{distribution generator functions}

  \item{name}{name of parameter}
}
\description{
Plot the weighted parameter values with underlying
distribution.
}

