% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict_GE}
\alias{predict_GE}
\title{Predicts galactose elimination (GE) for given vectors of liver volume and blood flow
based on GE response functions.}
\usage{
predict_GE(GEC_f, volLiver, flowLiver, galactose = 8, age = 20)
}
\description{
Under high galactose concentrations, i.e. gal=8.0mM, the maximal galactose elimination
rate is reached (GEC).
}

