% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEC_functions.R
\name{predict_GE_people}
\alias{predict_GE_people}
\title{Predict GE and GEkg for multiple people.}
\usage{
predict_GE_people(people, f_GE, volLiver, flowLiver, out_dir, galactose = 8)
}
\description{
Predicts the GEC and GECkg for given people. Calculates the
quantiles and stores some of the individual samples
}

