% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict_GEC}
\alias{predict_GEC}
\title{Predicts galactose elimination (GE) for given vectors of liver volume and blood flow
based on GE response functions.}
\usage{
predict_GEC(f_GE, volLiver, flowLiver, ages)
}
\description{
Under high galactose concentrations, i.e. gal=8.0mM, the maximal galactose elimination
rate is reached (GEC).
}

