% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict_liver_people}
\alias{predict_liver_people}
\title{Predict liver volume and bloodflow for people.}
\usage{
predict_liver_people(people, Nsample, sex_split = FALSE, Ncores = 1,
  debug = TRUE)
}
\description{
Predict liver volume and bloodflow for people.
}

