% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GAMLSS_functions.R
\name{predict_liver_person.fast}
\alias{predict_liver_person.fast}
\title{Predict liver volume and bloodflow for person.}
\usage{
predict_liver_person.fast(person, Nsample, sex_split)
}
\description{
Predict liver volume and bloodflow for person.
}

