% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict_liver_person}
\alias{predict_liver_person}
\title{Predict liver volume and bloodflow for person.
! Use the fast version whenever possible !}
\usage{
predict_liver_person(person, Nsample)
}
\description{
Predict liver volume and bloodflow for person.
! Use the fast version whenever possible !
}

