% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict_liverkg_people}
\alias{predict_liverkg_people}
\title{Predict liver volume per bodyweight and bloodflow per bodyweight for people.}
\usage{
predict_liverkg_people(people, Nsample, Ncores = 1, debug = TRUE)
}
\description{
Predict liver volume per bodyweight and bloodflow per bodyweight for people.
}

