% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict_liverkg_person.fast}
\alias{predict_liverkg_person.fast}
\title{Predict liver volume per bodyweight and bloodflow per bodyweight for person.}
\usage{
predict_liverkg_person.fast(person, Nsample, sex_split)
}
\description{
Predict liver volume per bodyweight and bloodflow per bodyweight for person.
}

