% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepare_fds}
\alias{prepare_fds}
\title{Handle not available distributions.}
\usage{
prepare_fds(f_ds, debug = TRUE)
}
\description{
Some of the distributons are not available due to missing
anthropomorhic information. Handle these cases so that
the combined probability distributions can be calculated correctly.
}

