% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{preprocess.ids}
\alias{preprocess.ids}
\title{Timecourses for ids loaded from the full set of data via preprocessing
The ids have to be part of the dictionary of the available names, which
is available in the SBML or via the CSV/Rdata.
fname <- getSimulationFileFromSimulationId(ma.settings$dir.simdata, simIds[1])
ids.dict <- names(data)
In standard analysis only the periportal and perivenious concentrations are used}
\format{\preformatted{ chr [1:12] "PP__alb" "PP__gal" "PP__galM" "PP__h2oM" "PP__rbcM" ...
}}
\usage{
preprocess.ids
}
\description{
Timecourses for ids loaded from the full set of data via preprocessing
The ids have to be part of the dictionary of the available names, which
is available in the SBML or via the CSV/Rdata.
fname <- getSimulationFileFromSimulationId(ma.settings$dir.simdata, simIds[1])
ids.dict <- names(data)
In standard analysis only the periportal and perivenious concentrations are used
}
\keyword{datasets}

