\name{preprocess}
\alias{preprocess}
\title{Get the data storage filename from the parameter file name.}
\usage{
preprocess(parsfile, sim.dir, outFile = NULL, max_index = -1)
}
\arguments{
  \item{parsfile}{Parameter file}

  \item{sim.dir}{directory with simulation data}
}
\value{
storage filename
}
\description{
Get the data storage filename from the parameter file name.
}

