% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preprocess_task}
\alias{preprocess_task}
\title{Preprocess the results for given folder.}
\usage{
preprocess_task(folder, ids = preprocess.ids, force = FALSE,
  out_name = "x")
}
\description{
Generates the combined Rdata file for the selected ids.
The collection of data files to be preprocessed is defined via
Read the timecourse data (CSV) and creates reduced data
structures consisting of some components of the full
sinusoidal system.
All simulation csv have been collected in a common folder.
Main challenge is the combination of the timecourses based
the varying timesteps used for simulation.
Important part is also dimension reduction of the data
structure.
Folder format follows '2014-11-17_T5'.
}

