% Generated by roxygen2 (4.0.2): do not edit by hand
\name{process_data_and_save}
\alias{process_data_and_save}
\title{Perform common processing steps and save the raw data.}
\usage{
process_data_and_save(data, dtype)
}
\description{
Sex/Gender is unified.
Set the data type (either individual or population)
Raw data is saved.
}

