% Generated by roxygen2 (4.0.2): do not edit by hand
\name{q_f}
\alias{q_f}
\title{Flow through pores function q(x).}
\usage{
q_f(x, p)
}
\description{
Solutions of the partial differential equations
q(x) flow through pores [m^2/s]
Volume flow results via scaling with length of capillary element.
}

