% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MultipleIndicator_functions.R
\name{rbc_time_offset}
\alias{rbc_time_offset}
\title{Calculation of time offset for experimental dilution curves
based on RBC peak.}
\usage{
rbc_time_offset(data)
}
\description{
Extrapolated time of first appearance of RBC tracer.
extrapolate between second and first point to zero
}

