% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readDataForSimulation}
\alias{readDataForSimulation}
\title{Load the column data for single simulation by sim id}
\usage{
readDataForSimulation(dir, simId, col.indices_f = NULL)
}
\arguments{
\item{sim}{simulation identifier}

\item{withTime}{keeps the time column}
}
\value{
column data
}
\description{
Load the column data for single simulation by sim id
}

