% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readDataForSimulationFile}
\alias{readDataForSimulationFile}
\title{Load the column data for simulation file.}
\usage{
readDataForSimulationFile(fname, col.indices_f = NULL)
}
\arguments{
\item{fname}{CSV file to load}
}
\value{
column data
}
\description{
Load the column data for simulation file.
}

