\name{readPPPVData}
\alias{readPPPVData}
\title{Read the data in a list structure
Parameter data pars has to be available in environment.}
\usage{
readPPPVData(max_index = -1)
}
\arguments{
  \item{max_index}{maximal index until which data is read}
}
\value{
list of PPPV data matrices
}
\description{
Read the data in a list structure Parameter data pars has
to be available in environment.
}

