% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{save_correlation_data}
\alias{save_correlation_data}
\title{Saves correlation data as csv and Rdata.}
\usage{
save_correlation_data(data, dir = get_correlation_dir())
}
\description{
Saves correlation data as csv and Rdata.
}

