% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save_shiny_fit_models}
\alias{save_shiny_fit_models}
\title{Save shiny fit models.}
\usage{
save_shiny_fit_models(dir = ma.settings$dir.shiny,
  file = file.path(ma.settings$dir.shiny, "fit_models.Rdata"))
}
\description{
Save the subset of GAMLSS models necessary for the individual predictions of
galactose elimination capacity.
}

