% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveRawData}
\alias{saveRawData}
\title{Saves the raw data of single source.}
\usage{
saveRawData(data, name, dir = getRawDir())
}
\description{
This saves the the experimental datasets after minimal preprocessing
and unification.
example: saveRawData('duc1979')
}

