% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{startDevPlot}
\alias{startDevPlot}
\title{Start a general plot.}
\usage{
startDevPlot(width = 1000, height = 1000, file = NULL, create_plots = F,
  res = 150)
}
\description{
Start a general plot.
}

