\name{stdlog}
\alias{stdlog}
\title{Calculate the stdlog for log-normal distribution from given mean and std}
\usage{
stdlog(m, std)
}
\arguments{
  \item{m}{mean}

  \item{std}{standard deviation}
}
\value{
stdlog for corresponding log-normal distribution
}
\description{
Calculate the stdlog for log-normal distribution from given
mean and std
}

