% Generated by roxygen2 (4.0.2): do not edit by hand
\name{storeFitData}
\alias{storeFitData}
\title{Writes the fit parameter to the global parameter object.}
\usage{
storeFitData(p.gen, fit, name)
}
\arguments{
\item{fit}{fit results}

\item{name}{name of the parameter}
}
\description{
Writes the fit parameter to the global parameter object.
}

