% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEC_functions.R
\name{subset_GE}
\alias{subset_GE}
\title{Create subset for GE response curves from integrated data.frame.}
\usage{
subset_GE(df)
}
\description{
Create subset for GE response curves from integrated data.frame.
}

