% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/string_tools.R
\name{trim}
\alias{trim}
\title{Returns string w/o leading or trailing whitespace}
\usage{
trim(x)
}
\arguments{
\item{x}{string to strip whitspaces}
}
\value{
stripped string
}
\description{
Returns string w/o leading or trailing whitespace
}

