% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wt.mean}
\alias{wt.mean}
\title{Weighted mean
x and wt have to be of same length
wt.mean <- sum(wt*x)/sum(wt)}
\usage{
wt.mean(x, wt)
}
\arguments{
\item{x}{data}

\item{wt}{weights}
}
\description{
Weighted mean
x and wt have to be of same length
wt.mean <- sum(wt*x)/sum(wt)
}

