% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wt.sd}
\alias{wt.sd}
\title{Weighted unbiased standard deviation
x and wt have to be of same length}
\usage{
wt.sd(x, wt)
}
\arguments{
\item{x}{data}

\item{wt}{weights}
}
\description{
Weighted unbiased standard deviation
x and wt have to be of same length
}

