% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wt.var}
\alias{wt.var}
\title{Weighted unbiased variance, i.e. (N-1)
x and wt have to be of same length}
\usage{
wt.var(x, wt)
}
\arguments{
\item{x}{data}

\item{wt}{weights}
}
\description{
Weighted unbiased variance, i.e. (N-1)
x and wt have to be of same length
}

