% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xlabByName}
\alias{xlabByName}
\title{Get the axis label.}
\usage{
xlabByName(p.gen, name)
}
\arguments{
\item{p.gen}{information about the parameters}

\item{name}{of the parameter}
}
\value{
axis label
}
\description{
Get the axis label.
}

