% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEC_functions.R
\name{xyz2matrix}
\alias{xyz2matrix}
\title{X,Y,Z data to matrix for bicubic sline interpolation}
\usage{
xyz2matrix(x, y, z, tol = 1e-08)
}
\description{
X,Y,Z data to matrix for bicubic sline interpolation
}

