% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.alpha}
\alias{add.alpha}
\title{Adds alpha value to a color.}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{col}{color}

\item{alpha}{alpha value between 0 and 1}
}
\value{
color with changed alpha value
}
\description{
Adds alpha value to a color.
}

