\name{addEventPoints}
\alias{addEventPoints}
\title{Handle Event timepoints.}
\usage{
addEventPoints(df)
}
\description{
Timepoints with corresponding data points are added for the
events. This is mainly for plotting the solution, which
would not represent correctly the fast change in components
at the events without these datapoints.
}

