% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{addMeanPopulationData}
\alias{addMeanPopulationData}
\title{Add data from population studies to dataset.}
\usage{
addMeanPopulationData(data, newdata)
}
\description{
I.e. multiple individuals put together
providing n (number subjects), mean (mean value) and Sd (standard deviation)
or range (distance to upper/lower limit of group).
Add n measurements of mean data.
}

