% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateProbabilitiesForVariables}
\alias{calculateProbabilitiesForVariables}
\title{Calculate the single variable probabilies based on ECDF}
\usage{
calculateProbabilitiesForVariables(pars, ecdf.list)
}
\description{
Calculate the single variable probabilies based on ECDF
}

