% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{calculateSampleProbability}
\alias{calculateSampleProbability}
\title{Calculate the overall probability of the sample under the assumption
of statistical independence.}
\usage{
calculateSampleProbability(pars, ps.var)
}
\description{
TODO: make sure it is valid
}

