\name{correctDilutionTimes}
\alias{correctDilutionTimes}
\title{Corrects the dilution times, so that zero timepoint is first datapoint.}
\usage{
correctDilutionTimes(data)
}
\arguments{
  \item{data}{data to be corrected}
}
\value{
new data with corrected times
}
\description{
Corrects the dilution times, so that zero timepoint is
first datapoint.
}

