% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MultiscaleAnalysis.R
\name{create.settings}
\alias{create.settings}
\title{Create the settings structure.}
\usage{
create.settings(dir = "/home/mkoenig", simulator = "ROADRUNNER")
}
\description{
Encodes the main dirs and settings necessary.
Simulator can be either ROADRUNNER or COPASI.
}

