\name{createColumnDataFiles}
\alias{createColumnDataFiles}
\title{Read all col.indices components and store R data structures.
For all simulations the csv files are loaded and the
respective columns extracted.}
\usage{
createColumnDataFiles(pars, dir, col.indices_f = NULL)
}
\arguments{
  \item{max_index}{maximal index until which data is read}
}
\value{
list of matrices
}
\description{
Read all col.indices components and store R data
structures. For all simulations the csv files are loaded
and the respective columns extracted.
}

