\name{createDataFromHistogramm}
\alias{createDataFromHistogramm}
\title{Generate data vector from histogramm.
Count data points are generated in the middle of every bin.}
\usage{
createDataFromHistogramm(dset)
}
\arguments{
  \item{dset}{histogram dataset}
}
\value{
data frame with x as data
}
\description{
Generate data vector from histogramm. Count data points are
generated in the middle of every bin.
}

