% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createDataMatrices}
\alias{createDataMatrices}
\title{Convert timecourse list structure into data matrix.
TODO: fix the performance (select only subset of interest)}
\usage{
createDataMatrices(ids, out.fname, simIds)
}
\arguments{
\item{datalist}{list of data matrices}
}
\value{
matrix of pppv data
}
\description{
Convert timecourse list structure into data matrix.
TODO: fix the performance (select only subset of interest)
}

