\name{createDataMatricesVarSteps}
\alias{createDataMatricesVarSteps}
\title{Convert timecourse list structure into data matrix.}
\usage{
createDataMatricesVarSteps(dir, datalist, time)
}
\arguments{
  \item{datalist}{list of data matrices}

  \item{compounds}{which compounds to take}
}
\value{
matrix of pppv data
}
\description{
Convert timecourse list structure into data matrix.
}

