% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{createListOfStandardECDF}
\alias{createListOfStandardECDF}
\title{Calculated ECDFs for standard distributions.}
\usage{
createListOfStandardECDF(p.gen, ps.var, Npoints = 25000)
}
\description{
Calculated ECDFs for standard distributions.
}

