% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extend_with_galactose_clearance}
\alias{extend_with_galactose_clearance}
\title{Extends the given parameters withthe clearance parameters and data.frame.}
\usage{
extend_with_galactose_clearance(processed, t_peak, t_end)
}
\description{
Necessary to provide the peak and simulation end time for calculation.
Here additional values from the simulation curves are added.
}

