% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEC_functions.R
\name{extend_with_galactose_clearance}
\alias{extend_with_galactose_clearance}
\title{Calculate clearance information for individual sinusoidal units.}
\usage{
extend_with_galactose_clearance(processed, t_end)
}
\description{
Galactose elimination (R), Clearance (CL), extraction ratio (ER) and
galactose difference (DG=ci-co) are calculated for all sinusoidal units
individually.
For the calculation the timecourse matrix is interpolated to get the
steady state value at exactly t_end. This is necessary due to the variable
step sizes of the integration (different step sizes for all solutions).
With the steady state concentrations of periportal galactose (PP__gal) and
perivenous galactose (PV__gal) the elimination values are calculated.
}

