% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extendParameterStructure}
\alias{extendParameterStructure}
\title{Create extended data frame with the calculated values based
on the information from the SBML.}
\usage{
extendParameterStructure(pars, fixed_ps, model)
}
\arguments{
\item{pars}{parameter structure}
}
\description{
Create extended data frame with the calculated values based
on the information from the SBML.
}

