% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBML_tools.R
\name{extendParameterStructure}
\alias{extendParameterStructure}
\title{Extend paramter structure with SBML information.}
\usage{
extendParameterStructure(pars, fixed_ps, model)
}
\arguments{
\item{pars}{parameter structure}

\item{fixed_ps}{fixed parameter names}

\item{model}{SBML model structure}
}
\value{
extended parameter structure
}
\description{
Create extended data frame with the calculated values based
on the information from the SBML.
The formulas for calculation are hard coded and have to be
updated with the respective model version.
}

