% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GAMLSS_functions.R
\name{f_d.factory.bodyweight}
\alias{f_d.factory.bodyweight}
\title{Density function for GAMLSS parameters (per bodyweight densities -> densities)}
\usage{
f_d.factory.bodyweight(pars)
}
\description{
Density function for GAMLSS parameters (per bodyweight densities -> densities)
}

