% Generated by roxygen2 (4.0.2): do not edit by hand
\name{f_d.flowLiverkg.pars}
\alias{f_d.flowLiverkg.pars}
\title{Calculate the liver bloodflow per bodyweight distributions for person.}
\usage{
f_d.flowLiverkg.pars(person)
}
\description{
Calculate the liver bloodflow per bodyweight distributions for person.
}

