% Generated by roxygen2 (4.0.2): do not edit by hand
\name{f_d.rejection_sample}
\alias{f_d.rejection_sample}
\title{Rejection sampling from given distribution.}
\usage{
f_d.rejection_sample(f_d, Nsim, interval)
}
\description{
Sampling from distributions via rejection sampling of good guess of
probability distribution.
}

