% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GAMLSS_functions.R
\name{f_d.volLiverkg.pars}
\alias{f_d.volLiverkg.pars}
\title{Calculate the liver volume per bodyweight distributions for person.}
\usage{
f_d.volLiverkg.pars(person)
}
\description{
Calculate the liver volume per bodyweight distributions for person.
}

