% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gender.cols}
\alias{gender.cols}
\title{Returns the colors for the different gender settings in data.}
\usage{
gender.cols()
}
\value{
gender levels and colors
}
\description{
Returns the colors for the different gender settings in data.
}

