% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateLogStandardParameters}
\alias{generateLogStandardParameters}
\title{Generates the standard parameters for meanlog and stdlog based on mean and std.}
\usage{
generateLogStandardParameters(p.exp)
}
\description{
The parameters are the reported mean and standard deviations from experiments.
These data is used if no histograms are available to fit the distributions.
Parameters are fitted in SI units.
}

