% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateLogStandardParameters}
\alias{generateLogStandardParameters}
\title{Generates the standard parameters for meanlog and stdlog based on mean and std.}
\usage{
generateLogStandardParameters()
}
\value{
data.frame of standard parameters
}
\description{
Generates the standard parameters for meanlog and stdlog based on mean and std.
}

