% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAllSBMLParameterIdsFromModel}
\alias{getAllSBMLParameterIdsFromModel}
\title{Parameter ids in the sbml model}
\usage{
getAllSBMLParameterIdsFromModel(model)
}
\arguments{
\item{filename}{SBML filename}
}
\description{
Parameter ids in the sbml model
}

