% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{getBreakPointsFromMidpoints}
\alias{getBreakPointsFromMidpoints}
\title{Calculatesbreakpoints for histogram data with equidistant midpoints.}
\usage{
getBreakPointsFromMidpoints(midpoints)
}
\arguments{
\item{midpoints}{vector of midpoints of bins}
}
\value{
vector of breakpoints
}
\description{
Calculatesbreakpoints for histogram data with equidistant midpoints.
}

