% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getColumnIndicesPPPV}
\alias{getColumnIndicesPPPV}
\title{Function to get the PPPV column indices from the data.}
\usage{
getColumnIndicesPPPV(data)
}
\arguments{
\item{data}{CSV data frame}

\item{withTime}{the time vector is part of the columns}
}
\value{
the column indices which should be used
}
\description{
Function to get the PPPV column indices from the data.
}

