% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFixedParameters}
\alias{getFixedParameters}
\title{Parameters which are fixed in the model}
\usage{
getFixedParameters(pars, all_ps)
}
\arguments{
\item{pars}{parameter structure}

\item{all_ps}{parameters in model}
}
\description{
Parameters which are fixed in the model
}

