\name{getProbabilitiesForSamples}
\alias{getProbabilitiesForSamples}
\title{Calculates the probability associated with the sample based on the
underlying distribution.}
\usage{
getProbabilitiesForSamples(pars, p.gen, name)
}
\arguments{
  \item{pars}{parameter structure}

  \item{p.gen}{distribution generator functions}

  \item{name}{name of parameter}
}
\description{
In which units is it coming out (transformed units) !
Important to keep track of the units ! TODO: do the fit for
unscaled data in basis units TODO: generalize to arbitrary
distributions of parameters
}

