% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSimulationFileFromSimulationId}
\alias{getSimulationFileFromSimulationId}
\title{Get integration timecourse file for simulation id.}
\usage{
getSimulationFileFromSimulationId(dir, simId, modelId)
}
\arguments{
\item{simId}{simulation identifier}
}
\value{
filename for the integration file
}
\description{
TODO: this is not working any more with the multiple simulations
}

