% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POPULATION_predict_functions.R
\name{individual_GEC_plot}
\alias{individual_GEC_plot}
\title{Plot individual GEC prediction.}
\usage{
individual_GEC_plot(person, data)
}
\description{
Plot individual GEC prediction.
}

