% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEC_settings.R
\name{integrate_GE_folders}
\alias{integrate_GE_folders}
\title{Preprocess set of data.frame of GE folders.}
\usage{
integrate_GE_folders(df_folders, factors = c("f_flow", "gal_challenge"),
  force = FALSE, t_end = 10000, f_tissue = 0.85, vol_liver = 1500)
}
\description{
Preprocess set of data.frame of GE folders.
}

