% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{loadParameterFile}
\alias{loadParameterFile}
\title{Read parameter file into data.frame.}
\usage{
loadParameterFile(file)
}
\arguments{
\item{file}{parameter csv file for loading}
}
\value{
data.frame of parameters
}
\description{
This function reads the parameter file in CSV format into a data.frame.
}

