% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadParameterFile}
\alias{loadParameterFile}
\title{Reads the stored parameter file as data.frame.}
\usage{
loadParameterFile(file)
}
\arguments{
\item{file}{parameter csv file for loading}
}
\value{
data.frame of parameters
}
\description{
This function reads the parameter file in CSV format into a data.frame.
 To load the associated data for the simulations do the preprocessing and
 use the load functions.
}

