% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MultiscaleAnalysis.R
\docType{data}
\name{ma.settings}
\alias{ma.settings}
\title{Settings structure with available directories.}
\format{\preformatted{List of 8
 $ dir.base   : chr "/home/mkoenig/multiscale-galactose"
 $ dir.code   : chr "/home/mkoenig/multiscale-galactose/R"
 $ dir.expdata: chr "/home/mkoenig/multiscale-galactose/experimental_data"
 $ dir.simdata: chr "/home/mkoenig/multiscale-galactose-results"
 $ dir.results: chr "/home/mkoenig/multiscale-galactose-results"
 $ dir.gamlss : chr "/home/mkoenig/multiscale-galactose/results/gamlss"
 $ dir.shiny  : chr "/home/mkoenig/multiscale-galactose/R/shiny/gec_app/data/gamlss"
 $ simulator  : chr "ROADRUNNER"
}}
\usage{
ma.settings
}
\description{
Settings structure with available directories.
}
\keyword{datasets}

