\name{meanlog}
\alias{meanlog}
\title{Calculate the meanlog for log-normal distribution from given mean and std}
\usage{
meanlog(m, std)
}
\arguments{
  \item{m}{mean}

  \item{std}{standard deviation}
}
\value{
meanlog for corresponding log-normal distribution
}
\description{
Calculate the meanlog for log-normal distribution from
given mean and std
}

