% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{MultiscaleAnalysis}
\alias{MultiscaleAnalysis}
\alias{MultiscaleAnalysis-package}
\title{MultiscaleAnalysis: Supporting package for multiscale-galactose model.}
\description{
The MultiscaleAnalysis package is a set of tools that implement the
necessary R analysis functions for the SBML model and simulations of a
liver multiscale model developed in the Virtual Liver Network (VLN).
}
\section{Introduction}{


Main parts of the model are the visualization and integration of sets of simulations,
and the prediction of Galactose Elimination Capacity (GEC) for individuals based on
their anthropomorhpic information.
}

