\name{outfileFromParsfile}
\alias{outfileFromParsfile}
\title{Get the data storage filename from the parameter file name.}
\usage{
outfileFromParsfile(parsfile)
}
\arguments{
  \item{parsfile}{Parameter file}
}
\value{
storage filename
}
\description{
Get the data storage filename from the parameter file name.
}

