\name{outfileFromParsFile}
\alias{outfileFromParsFile}
\title{Get the data storage filename from the parameter file name.}
\usage{
outfileFromParsFile(parsfile)
}
\arguments{
  \item{parsfile}{Parameter file}
}
\value{
storage filename
}
\description{
Get the data storage filename from the parameter file name.
}

