\name{plotDilutionData}
\alias{plotDilutionData}
\title{Plot one multiple-dilution indicator dataset}
\usage{
plotDilutionData(data, compounds = compounds, ccolors = ccolors,
  correctTime = FALSE)
}
\arguments{
  \item{data}{dataset to be plotted}

  \item{correctTime}{should the time be corrected}
}
\description{
Plot one multiple-dilution indicator dataset
}

