% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDilutionData}
\alias{plotDilutionData}
\title{Plot single multiple-dilution indicator dataset.}
\usage{
plotDilutionData(data, compounds, ccolors, correctTime = FALSE)
}
\arguments{
\item{data}{dataset to be plotted}

\item{correctTime}{set TRUE if the time should be corrected}
}
\description{
Plot single multiple-dilution indicator dataset.
}

