\name{plotDilutionData}
\alias{plotDilutionData}
\title{Plot single multiple-dilution indicator dataset.}
\usage{
plotDilutionData(data, compounds = compounds, ccolors = ccolors,
  correctTime = FALSE)
}
\arguments{
  \item{data}{dataset to be plotted}

  \item{correctTime}{set TRUE if the time should be
  corrected}
}
\description{
Plot single multiple-dilution indicator dataset.
}

