\name{plotLogNormalDistribution}
\alias{plotLogNormalDistribution}
\title{Plot the density distribution.}
\usage{
plotLogNormalDistribution(p.gen, name, maxvalue, N = 1000, lcolor = "blue")
}
\arguments{
  \item{p.gen}{information about the parameters}

  \item{name}{of the parameter}

  \item{maxvalue}{for calculation of distribution}
}
\description{
Plot the density distribution.
}

