% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{plotLogNormalDistribution}
\alias{plotLogNormalDistribution}
\title{Plot the log normal density distribution.}
\usage{
plotLogNormalDistribution(p.gen, name, max.value, N = 1000, lcolor = "blue")
}
\arguments{
\item{p.gen}{information about the parameters}

\item{name}{of the parameter}

\item{max.value}{for calculation of distribution}
}
\description{
Plot the log normal density distribution.
}

