% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotLogNormalDistribution}
\alias{plotLogNormalDistribution}
\title{Plot the log normal density distribution.}
\usage{
plotLogNormalDistribution(p.gen, name, maxvalue, N = 1000, lcolor = "blue")
}
\arguments{
\item{p.gen}{information about the parameters}

\item{name}{of the parameter}

\item{maxvalue}{for calculation of distribution}
}
\description{
Plot the log normal density distribution.
}

