% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{plotParameterHistogram}
\alias{plotParameterHistogram}
\title{Plot parameter histogram}
\usage{
plotParameterHistogram(pars, name, breaks = 40)
}
\description{
Plot parameter histogram
}

