% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotParameterHistogram}
\alias{plotParameterHistogram}
\title{Plot parameter histogram}
\usage{
plotParameterHistogram(p.gen, name)

plotParameterHistogram(p.gen, name)
}
\arguments{
\item{name}{of parameter}

\item{pars}{Parameter data frame}
}
\value{
hist information
}
\description{
Plot parameter histogram

Plot parameter histogramm of the sinusoidal unit samples.
}

