% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{plotParameterHistogramFull}
\alias{plotParameterHistogramFull}
\title{Creates histogram for all parameters.}
\usage{
plotParameterHistogramFull(pars, all_pars = FALSE)
}
\arguments{
\item{pars}{parameter data structure.}
}
\description{
This function generates all the histograms for the parameters in the
given parameter structure.
}

