% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotProbabilitiesForVariable}
\alias{plotProbabilitiesForVariable}
\title{Plot probabilities with associated midpoints as horizontal lines}
\usage{
plotProbabilitiesForVariable(x, f.ecdf, xlab)
}
\description{
Plot probabilities with associated midpoints as horizontal lines
}

