% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict_GEkg}
\alias{predict_GEkg}
\title{Predicts galactose elimination per bodyweight (GEkg)
for given vectors of liver volume and blood flow based on GE response functions.}
\usage{
predict_GEkg(f_GE, volLiverkg, flowLiverkg, ages)
}
\description{
Under high galactose concentrations, i.e. gal=8.0mM, the maximal galactose elimination
rate is reached (GEC).
}

