% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEC_functions.R
\name{predict_GEkg}
\alias{predict_GEkg}
\title{Predicts galactose elimination per bodyweight (GEkg)
for given vectors of liver volume and blood flow based on GE response functions.}
\usage{
predict_GEkg(f_GE, volLiverkg, flowLiverkg, galactose = 8, age = 20)
}
\description{
Under high galactose concentrations, i.e. gal=8.0mM, the maximal galactose elimination
rate is reached (GEC).
}

