% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict_liver_person.fast}
\alias{predict_liver_person.fast}
\title{Predict liver volume and bloodflow for person.}
\usage{
predict_liver_person.fast(person, Nsample, sex_split)
}
\description{
Predict liver volume and bloodflow for person.
}

