% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GAMLSS_functions.R
\name{prepare_fds}
\alias{prepare_fds}
\title{Handle not available distributions.}
\usage{
prepare_fds(f_ds)
}
\description{
Some of the distributons are not available due to missing
anthropomorhic information. Handle these cases so that
the combined probability distributions can be calculated correctly.
}

