% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preprocess_task}
\alias{preprocess_task}
\title{Preprocess the results for given task.}
\usage{
preprocess_task(folder, ids = preprocess.ids, force = FALSE)
}
\description{
Generates the combined Rdata file for the selected ids.
The collection of data files to be preprocessed is defined via
}

