\name{readDataForSimulation}
\alias{readDataForSimulation}
\title{Load the column data for single simulation by sim name}
\usage{
readDataForSimulation(dir, simId, col.indices_f)
}
\arguments{
  \item{sim}{simulation identifier}

  \item{withTime}{keeps the time column}
}
\value{
column data
}
\description{
Load the column data for single simulation by sim name
}

