% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GAMLSS_predict_functions.R
\name{save_shiny_fit_models}
\alias{save_shiny_fit_models}
\title{Save shiny fit models.}
\usage{
save_shiny_fit_models(dir = file.path(ma.settings$dir.base, "results",
  "gamlss"), file = file.path(ma.settings$dir.code, "shiny", "gec_app",
  "data", "gamlss", "fit_models.Rdata"))
}
\description{
Save shiny fit models.
}

