% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trim.trailing}
\alias{trim.trailing}
\title{Returns string w/o trailing whitespace}
\usage{
trim.trailing(x)
}
\arguments{
\item{x}{string to strip whitspaces}
}
\value{
stripped string
}
\description{
Returns string w/o trailing whitespace
}

