% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{ylabByName}
\alias{ylabByName}
\title{Get the axis label.}
\usage{
ylabByName(p.gen, name)
}
\arguments{
\item{p.gen}{information about the parameters}

\item{name}{of the parameter}
}
\value{
axis label
}
\description{
Get the axis label.
}

